Refactor: more modular split across source files
