--- src/evaluating_rewards/analysis/plot_all_gridworld_reward.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-CONFIGS="sparse_goal sparse_goal_shift sparse_goal_scale \
-         dense_goal antidense_goal transformed_goal center_goal \
-         dirt_path cliff_walk sparse_penalty all_zero"
-
-parallel --header : python -m evaluating_rewards.analysis.plot_gridworld_reward \
-          with {config} \
-          ::: config ${CONFIGS}
