+++ src/evaluating_rewards/analysis/reward_figures/gridworld_reward_heatmap.py
